Allow GLIFTOPT max glift to be exceeded by epsilon.
Allow GLIFTOPT item 2 "Maximum gas lift supply limit" to be exceeded by
a small value epsilon (a fraction of the increment size). This was done
to work around the case where the max limit is a multiple of the increment
and rounding errors makes a group's ALQ slightly higher than the
maximum limit preventing it from reaching the maximum limit.