Fixed logic when bhp cannot be determined from thp.
When gas lift is switched off by setting ALQ to zero, and later
in the schedule is switched back on again, it might not be possible to
determine bhp from thp for low small ALQ values. Instead of aborting the
gas lift optimization, we should try increasing ALQ until we get
convergence or until the maximum ALQ for the well is reached.