Test with two python simulators modules
Test if we can work around the issue with static variables in the
opmsimulators library not being reset between python test cases.
This commit creates two independent Python modules simulators.so and
simulators2.so and loads each of them in two different test cases
in the same test file test_basic.py

Still this does not work. The first test case runs, but when starting the
second test case, it aborts with:

"The MPI_Init() function was called after MPI_FINALIZE was invoked."

Could this be a problem with static variables in libmpi.so? Since both
simulators.so and simulators2.so link dynamically with libmpi.so ?