Refactor step()
Refactor step() to into a substepsLoop_() method that takes a sub step
timer as input. This is to prepare implementation of timestepping for
a slave reserovir in a coupled reservoir simulation. In this way, we can
later create a new stepSlave() method based on step() that can reuse
substepsLoop_() with a different sub step timer. This timer will do sub
stepping for timesteps submitted from a master process to a slave process
(which will generally be different from a report step)